[Oztechan/CCC#1421] Extract ServerController
