+++ backend/src/jvmMain/kotlin/com/oztechan/ccc/backend/controller/server/ServerController.kt
@@ -1,0 +1,5 @@
+package com.oztechan.ccc.backend.controller.server
+
+internal interface ServerController {
+    suspend fun getOfflineCurrencyResponseByBase(base: String): String?
+}
