+++ backend/src/jvmMain/kotlin/com/oztechan/ccc/backend/di/module/BackendModule.kt
@@ -1,7 +1,7 @@
 package com.oztechan.ccc.backend.di.module
 
-import com.oztechan.ccc.backend.di.module.submodule.repositoryModule
+import com.oztechan.ccc.backend.di.module.submodule.controllerModule
 
 internal val backendModules = buildList {
-    add(repositoryModule)
+    add(controllerModule)
 }
