+++ backend/src/jvmMain/kotlin/com/oztechan/ccc/backend/module/Server.kt
@@ -1,7 +1,7 @@
 package com.oztechan.ccc.backend.module
 
 import co.touchlab.kermit.Logger
-import com.oztechan.ccc.backend.repository.api.ApiRepository
+import com.oztechan.ccc.backend.controller.server.ServerController
 import com.oztechan.ccc.backend.routes.getCurrencyByBase
 import com.oztechan.ccc.backend.routes.getError
 import com.oztechan.ccc.backend.routes.getRoot
@@ -17,17 +17,17 @@ import org.koin.java.KoinJavaComponent.inject
 
 @Suppress("unused")
 internal fun Application.server() {
-    val apiRepository: ApiRepository by inject(ApiRepository::class.java)
+    Logger.i { "Application server" }
+
+    val serverController: ServerController by inject(ServerController::class.java)
     val globalScope: CoroutineScope by inject(CoroutineScope::class.java)
     val ioDispatcher: CoroutineDispatcher by inject(CoroutineDispatcher::class.java, named(DISPATCHER_IO))
 
     routing {
-        Logger.i { "start rooting" }
-
         globalScope.launch(ioDispatcher) {
             getError()
             getRoot()
-            getCurrencyByBase(apiRepository)
+            getCurrencyByBase(serverController)
             getVersion()
         }
     }
