--- backend/src/jvmMain/kotlin/com/oztechan/ccc/backend/repository/api/ApiRepository.kt
@@ -1,6 +1,0 @@
-package com.oztechan.ccc.backend.repository.api
-
-internal interface ApiRepository {
-    fun startSyncApi()
-    suspend fun getOfflineCurrencyResponseByBase(base: String): String?
-}
