+++ backend/src/jvmMain/kotlin/com/oztechan/ccc/backend/routes/CurrencyRoute.kt
@@ -5,7 +5,7 @@
 package com.oztechan.ccc.backend.routes
 
 import co.touchlab.kermit.Logger
-import com.oztechan.ccc.backend.repository.api.ApiRepository
+import com.oztechan.ccc.backend.controller.server.ServerController
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.application.call
 import io.ktor.server.response.respond
@@ -15,10 +15,10 @@ import io.ktor.server.routing.get
 private const val PATH_BY_BASE = "/currency/byBase/"
 private const val PARAMETER_BASE = "base"
 
-internal suspend fun Route.getCurrencyByBase(apiRepository: ApiRepository) = get(PATH_BY_BASE) {
+internal suspend fun Route.getCurrencyByBase(serverController: ServerController) = get(PATH_BY_BASE) {
     call.parameters[PARAMETER_BASE]?.let { base ->
         Logger.i { "GET Request $PARAMETER_BASE $base" }
-        apiRepository.getOfflineCurrencyResponseByBase(base)?.let {
+        serverController.getOfflineCurrencyResponseByBase(base)?.let {
             call.respond(it)
         } ?: call.respond(HttpStatusCode.NotFound)
     } ?: run {
